+++ lib/loader.rb
@@ -4,7 +4,7 @@ require_relative './domain/team'
 require_relative './use_cases/fetch_pull_requests'
 require_relative './use_cases/fetch_teams'
 require_relative './use_cases/send_slack_messages'
-require_relative './use_cases/view_cacher'
+require_relative './use_cases/cache'
 require_relative './gateways/pull_request'
 require_relative './gateways/slack_message'
 require_relative './gateways/team'
