+++ spec/use_cases/cache_spec.rb
@@ -1,16 +1,16 @@
-describe UseCases::ViewCacher do
+describe UseCases::Cache do
   context 'Given the result has not been cached' do
     it 'caches the result' do
       file = double(exists?: false)
       expect(file).to receive(:open)
-      described_class.new(cache_file: 'foo.html', file: file).execute do
+      described_class.new(name: 'foo', file: file).execute do
         'some cached result'
       end
     end
 
     it 'returns the result' do
       file = double(exists?: false, open: nil)
-      result = described_class.new(cache_file: 'foo.html', file: file).execute do
+      result = described_class.new(name: 'foo', file: file).execute do
         'some cached result'
       end
 
@@ -21,7 +21,7 @@ describe UseCases::ViewCacher do
   context 'Given the result has been previously cached' do
     it 'returns the cached result' do
       file = double(exists?: true, mtime: Time.now, read: 'foo')
-      result = described_class.new(cache_file: 'foo.html', file: file).execute { }
+      result = described_class.new(name: 'foo.html', file: file).execute { }
       expect(result).to eq('foo')
     end
   end
@@ -30,7 +30,7 @@ describe UseCases::ViewCacher do
     it 'populates the cache' do
       file = double(exists?: true, mtime: Time.now - 121, open: nil)
       expect(file).to receive(:open)
-      described_class.new(cache_file: 'foo.html', file: file).execute {}
+      described_class.new(name: 'foo', file: file).execute {}
     end
   end
 end
